[FIX] delivery_dhl: fix error when using shipping method with dutiable for domestic delivery - Install Sales (sale_management) and DHL Express Shipping (delivery_dhl) - Go to Sales > Configuration > Shipping Methods - Select a DHL delivery method (e.g. DHL USA) and activate "Dutiable Material" option - Go to Sales and create a quotation - Add a customer whose an addess is in the same country than the company - Add a product with a Weight and a Volume - Click on the "ADD SHIPPING" button - On the wizard, select the modified DHL delivery method and click on "GET RATE" button The following error occurs: "AttributeError: 'NoneType' object has no attribute 'QtdShp'"
When making a request for a domestic delivery with a shipping method with dutiable option,
DHL API returns a response with a success status but with no "success" data.
Instead there is a note in the response with error code 410301
and message "Products not available between this origin and destination (network segment)".
As Odoo receives a successful response, it tries to read the "success" data and crashes.

For this use case, we will catch the AttributeError and if the 410301 error code is in the notes of the response,
we will display the DHL error message to customer and a hint explaining that the issue could come from dutiable option.

opw-2263982

X-original-commit: 581fbf530b47b042a2cf2d8121f9af66a68f5648