[FIX] mrp_workorder: correct definition of 'is_first_started_wo'
Start a workorder without doing anything. Leave it. Start another
workorder and record production. We would like this second workorder
create the backorder but it's not the case as it's not the first started
workorder. Instead, we want it to be the first 'recorded' workorder.

Task : 2278147

closes odoo/enterprise#11645

X-original-commit: 9f6ba9c0bbd2982a91651a742ccce6000e6ae25b
Related: odoo/upgrade#1451
Related: odoo/odoo#54042
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>