[FIX] sign: Fix sign_widgets_tour random fail on runbot
Currently, The tour 'sign_widgets_tour' fails randomly when runned
multiple times(fail ratio 200:1 build). It shows a traceback with the
following message: "Uncaught TypeError: Cannot read property
'resetCanvas' of undefined"

The tour is failed because sometimes the dialog is destroyed before the
_printImage completes the process to draw the image. in _printImage
there is drawTimeout so it executes the process at last and before that
the tour clicks on 'Adopt and Sign' so dialog is going to destroy and
due to that issue is generated.

So in this commit, wait for 1 sec before click on 'Adopt and Sign' so
the process of _printImage is completed.

closes odoo/enterprise#11692

Taskid: 2245597
Closes: #11541
X-original-commit: 63df6b258bd5664ae5bf69513aedbae019755d62
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>