[FIX] hr_contract_salary: hidden field should not be required
Issue description:

On the salary configurator page, the field for the spouse information
were required even if the marrital status was single

The javascript that set the fields required or not according their
visibility status was not searching for the input properly.

Solution:

Search for the input to show/hide properly

UX improvement:
Scroll the user to the top of personnal information form
and not at the bottom.

closes odoo/enterprise#11516

X-original-commit: 7945b6a252231cc357b9aada9ecc85bda418b1fb
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>