[IMP] web_mobile: restore scroll position after closing a modal
Before, if any dialog is opened on mobile, the 'modal-open'
class is added to body and with that class, position of body
is set to fixed, and so body is always scrolled to the top.
As a consequence, user had to scroll again to the element s/he
was viewing after closing the dialog.

Now, we store the current scroll position when opening a dialog
to be able to restore the scroll position while closing the last
visible dialog.

Not that we had to do this "temporary" fix because the Chromium bug
is still not fixed.
More information in this original commit:
https://github.com/odoo/enterprise/commit/d878c2c605102e70ae9d710cf353cbb2d89e9482

Task Id: 2152130
PR #11841

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>