[FIX] l10n_mx_reports_closing: Avoid errors on trial balance
In the next case:
A company starts operations in 2018, in December makes the close
entry, and mark the entry with the option "MARK AS CLOSING ENTRY FOR THE
FISCAL YEAR".
Get the trial balance and is ok.
In 2019 makes the close entry with the same option.

But when in 2019, try to get the "Trial Balance Report", ignore the
close entry for 2018 moves.

That is incorrect because the entry in 2018 must be considered in the
new years.

OPW #2276153

closes odoo/enterprise#11646

X-original-commit: e9d5855e0e5035d8c491cbf7fa71d8a2276efbc9
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>