[FIX] helpdesk_(timesheet)_grid: Prevent the stoping timer when time dialog cancelled
Currently, When some one start the timer and click on stop button at
that time it was first stop the timer by super method and then open
the time dialog so timer vanishes and the Start button is displayed again.

After this commit, when click on stop button do not stop the timer by
calling super method of mixin just open the dialog with minutes spent
and on click of SAVE in dialog stop the time and unlink the related
timer entry and on click of CANCEL of dialog it will not stop the timer.

closes odoo/enterprise#12347

Task-id: 2230217
X-original-commit: 6e5e02cb31d3b30f7c83477ff908a634150f1fd2
Related: odoo/odoo#55744
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>