[FIX] delivery_dhl: fix xml parsing of rating response In Sales Quotation, when adding DHL as shipping method and computing rate, the following error may be triggered "AttributeError: 'NoneType' object has no attribute 'settings'".
It comes from the fact that sometimes, when requesting rate from DHL API,
its XML response does not respect its schema definition (XSD).
Which triggers an error in zeep library.

As DHL API responses do not seem to always respect its XSD, the parsing of the XML response for the rate
will be done with "lxml" instead of "zeep", as it was in v12.

opw-2263982

closes odoo/enterprise#11583

X-original-commit: bededf2e8bfb21c602e5ededabb4444a2d2527c7
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>