[IMP] documents: Documents search panel Owl
Adaptation of the documents extension of the search panel to the newly
refactored search panel and model.

The logic regarding the rendering and the interactions has been given to
the "Documents Search Panel" component while the data logic has shifted
to a new model extension: "Documents Search Panel Extension".

The other documents files have been adapted to reflect the changes on
the searchModel and the search panel component.

Part of task 2228968