[IMP] web_enterprise: Mobile search panel/search view Owl
Adaptation of the mobile extension of the search panel to the newly
refactored search panel and model and unification of the search views
layouts (control panel and search panel)

The mobile version of the search panel has been adapted to the
owlification of the search panel.

Also the search view layouts used by both the search panel and control
panel have been merged to increase consistency and maintainability.
This includes a few DOM modifications in the control panel mobile XML
and SCSS files as well as a major refactor of the mobile search panel
layout.

Part of task 2228968