[FIX] planning: gantt: expand and collapse special empty rows.
In the planning gantt view, special empty rows are sometimes created.
Before this commit, if the user would try to collapse one of these rows
playing the role of a group header, a crash would occur ("Cannot set
property 'isOpen' of undefined").

With this commit, this is no longer the case. Moreover, it is also
possible to expand such a group header: everything will work as
expected.

For this we had to
- add the empty rows in the object allRows (this is related to the bad
  assignment mentioned above),
- set adequately the keys childrenRowIds and groupedBy in some rows
  in order to be able to destroy/update correctly the gantt rows.

Task ID: 2283977

closes odoo/enterprise#12563

X-original-commit: dc0e171b5f2f560d66753e0367889f435ac1dc78
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>