[FIX] web_studio: export is_mail_activity field
- with studio create a new model;
- in the form view add the chatter widget;
- export the web_studio customizations;
- in a new database, import the customizations.

Before this commit, an error was raised when importing the
customizations: "Field `activity_ids` does not exist".
This error occurs because the is_mail_activity field of the original
model is not exported. When importing the customization into a new
database, the new model will have the is_mail_activity field set as
false by default.

Now, the field is_mail_activity is correctly exported, and the model and
views are correclty imported.

opw-2278509

closes odoo/enterprise#11698

X-original-commit: 5b626c1d69d1d355d8db3b5422cb7a34c38562f3
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>