[ADD] digest_enterprise: add a bridge module for enterprise digest tips
In this commit we add a new bridge module between digest and web_enterprise.
We would like to showcase web_enterprise features and therefore require a bridge module with digest.

We add a new tip: "Navigate menus like a pro”, used to showcase the enterprise-specific Home main menu.

Task ID : 2274264
PR #11685

This commit is a pre-merge in order to backport it in 13.4 along with other tips.
See odoo/odoo#53580 and odoo/enterprise#11394.
Migrations script will be completed with odoo/upgrade#1406.
This trick is necessary to ensure runbot coherency.

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>