[FIX] helpdesk: fix extra day in SLA deadline when ticket is created after working hours
- Install Helpdesk
- Confugure working hours with values that are earlier than current time for every working days
(For example, if the current time is 11:00, sets 03:00 - 10:00 for working hours)
- Go to Helpdesk > Configuration > SLA Policies and create a SLA Policy:
  * Team: "Customer Care"
  * Reach Stage: Solved
  * Reach In: 1 day 0 hour
- Go to Helpdesk > Configuration > Helpdesk Teams and configure "Customer Care" team:
  * Activate SLA Policies
  * Select the previously configured calendar with our working hours
- Go to Helpdesk > Overview and create a new ticket for "Customer Care"
The deadline is computed with 1 extra day.

The deadline is computed in the following way:
- If SLA Policy contains days,
  1) plan_days is executed with the SLA Policy days, returning the correct day for the deadline with the end time of the working day
  2) the time of the deadline is set to the time of the ticket creation
- In all cases,
  3) plan_hours is executed with the SLA Policy hours, returning the final deadline
    (adding SLA Policy hours and making sure that deadline is within working hours)
In our case, at the end of step 2, the correct time should be the start time of the working day.
Instead, we have the time of the ticket creation which is later than the end time of the working day.
And to finish, step 3 returns the first valid datetime, which is the start date of the following working day;
resulting on 1 extra day for the deadline.

To prevent this, the time is set to 0:00 in step 2 if the time of the ticket creation is later
than the end time of the working day.
And then, the deadline is set to the start time of the current working day in step 3.

opw-2287201

closes odoo/enterprise#11966

X-original-commit: 00a596dbef98626a8438c0b9c30b47398aece688
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>