[IMP] stock_barcode: add keyboard shortcuts for increment buttons
This commit adds in keyboard shortcuts using the letters a-z (for
incrementing 1 at a time) and A-Z (for incrementing remaining products
when applicable) in transfers when app is opened in a non-touchscreen
environment. These shortcuts are displayed underneath the corresponding
button they trigger. These shortcuts only work when 'ctrl', 'alt', and
'command' (Apple computers) are not also being pressed. When there are
more than 26 lines, then remaining lines do not have a shortcut key.

Also adds a settings option (that is active by default) for whether
increment buttons have these keyboard shortcuts and which keyboard
layout the letters should be arranged in (starting from top left), which
will be company dependent. Note this is in combination with a change in
the settings layout that affects community.

Shortcut keys are also not hidden when accesskeys are displayed due to
poor interactions with accesskey logic and how the web browser is
configured.

Part of specificiation "8. No touch: keyboard shortcuts" of overall
barcode improvement task.

Task: 1884520
Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086