[FIX] sale_subscription: wrong behavior during renewal
Issue:
1) During the renewal process we expect to see in the subscription
the same lines than on the renewal sale order.

It was not the case.
- If you remove lines from the subscription on the quotation,
  the lines remains on the subscription after the confirmation
- If you have one line for a given product on the subscription before the renewal
  you add a second line with the same product on the renewal quotation
  You end up with only one line on the subscription with the quantity of
  the last sale order line

2) Change in template_id of the subscription may change the MRR
of the subscription if the template has a different recurring rule.
This change was not tracked properly in the sale.subscription.log

Solution:
1) The subscription should be wiped at renewal, so the
confirmation of the SO will create all the line from the so
on the subscription. This way the subscription is a mirror
of the renewal quotation

2) Create log if template_id is in the tracking value

closes odoo/enterprise#13004

X-original-commit: 569d535f1c597d803d05b641c9c4b4007369877e
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>
Signed-off-by: Thibault Francois <tfr-odoo@users.noreply.github.com>