[FIX] sign: Fix PDF sizes computation
Before this commit:
The height and width of the PDF is calculated based on its upper right point in space.
The way to compute is correct, however some PDF may have incorrect meta-data that would lead to a miscalculation.
For instance `MediaBox[0 792 612 0]` will lead to a height of 0 despite the fact that it should be 792.
Meta-data wrong data is probably the consequence of the application that created them.

After this commit:
Use the more accurate `getWidth` and `getHeight` absolutes value (as it may be negative)

OPW-2280195

closes odoo/enterprise#11927

X-original-commit: f4d2651870c9d945d89e829efd34fc84ca552909
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>