[IMP] web_enterprise: enable action menu on mobile
This commit enables ControlPanel's ActionMenus ('Print' and 'Action') in
List/Form view on mobile.

The main purpose of enabling it is to give users access to the basic
actions but also the custom actions implemented in the client's
database.

The initial challenge of this implementation was the small space
available on mobile screens. To save space, we reused the selection
banner but moved it to the ControlPanel bottom section instead of hiding
the columns' names (as they are useful to the user). This implementation
UX also matches common patterns used by Material applications.

Also, still in the process of saving as much space as possible to
display more useful information instead to the user, the row's
checkboxes are now hidden on mobile and replaced by a 'highlight' of the
selected rows.

Note: we had to disable the "read-only editable list" as interacting
with selection AND keeping this feature was heavily error-prone for the
user.

Reference:
https://material.io/design/interaction/selection.html

Task ID: 1879566