[FIX] mrp_workorder: No record_production when workorder is done
When click the "mark as done and close mo" button, we will call
do_finish and record the production. If a comsumption warning wizard is
launched and we confirm it, we will call do_finish again, but no
production can be recorded and give a error. This commit check the state
in do_finish, if it's done, don't record the production.

Task 2278147

X-original-commit: 466599c07a5bf4d42773ca98b232ccfdbf8556a1