[FIX] mrp_workorder: fix action_assign in tests
action_assign was sometime called before confirming the MO meaning it
had no effect since the moves were still in draft state.
In addition to fixing this, as the quantities are now assigne correctly
we no longer need to manually set the lot_id and can make an additional
check for it.

X-original-commit: d14c544f03f9d066c78447fdf4b9eb5cb05e241d