[FIX] l10n_mx_edi: Portal view of invoices
- There was an error with the CSS rendering the QR code,
the CSS was forcing the dimensions of the QR code to
180x180 pixels but the call to the controller that renders
the QR code was returning an image of 600x100 pixels, that
style was stretching the image and the QR looked odd, in
order to fix it the dimensions of 180x180 were passed directly
to the controller and the style was removed, this way the
image is in native size.

- There was another error with the original string, it was being
chopped by an old style that was not necessary since the default
style for divs makes the automatic jumplines.

closes odoo/enterprise#11855

X-original-commit: 002d2eccd23d49c41950fbfb1c15dc34480bbed0
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>