[IMP] *: default_get overrides cleanup
1) Do not specify unwanted defaults in default_get overrides

When default_get is called, the wanted fields are specified through the
fields_list arg.  It is useless to fill the values for unwanted fields.

As default_get is called for nearly all records creation, simplifying
the default_get overrides:
* remove potential wrong side-effects of the values
* remove some useless or wrong  defaults computations (searches, refs,
...)

2) Replace default_get overrides by default attribute on fields when
possible

Default method on fields will only be called at the ORM level when
needed, and default values returned will be correctly post-processed
when needed.

3) Fix some some multi-company bugs (rental_sign, planning).