[IMP] hr_referral: Improve the empty list content in reward
Currently, In the referral module when the 'Reward','My Referral
and  job position kanban view is empty then it displays a smiling
face with help message.

In this commit, we replace a smiling face class with a points-gift
image for reward and my referral, refer-friends for job position
,improve the help message and adde help empty content for lisy and
enable sample data.

closes odoo/enterprise#11857

Taskid: 2297184
Closes: #11857
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>