[IMP] stock_barcode: show/hide shortcut keys based on shift
This commit makes it so only lowercase shortcut keys are shown in
barcode app increment buttons (when setting is active) "shift" is
not pushed and only uppercase shortcut keys are shown when "shift"
is pushed. There are bunch of unavoidable (or hugely annoying to handle)
situations because of this. Therefore we have the following assumptions:

- We can ignore if CAPS LOCK is active otherwise what is visible and
  what is expected to happen when a letter is pushed will be confusing.
- We can ignore if 2 shift buttons are pushed at the same time and only
  1 is unpushed (lower case will be displayed, but upper case button will
  be triggered if a letter is pushed).
- We can ignore if the user has the shift button pushed and switches
  windows/focus (i.e. upper case letters will be shown even if they won't
  necessarily be triggered).

Part of specificiation "8. No touch: keyboard shortcuts" of overall
barcode improvement task.

Task: 1884520
Community PR: odoo/odoo#55096
Closes: odoo/enterprise#12086