[FIX] social: emoji widget overlap the image picker
Bug
===
On the brave browser, under Windows, in the comment modal,
the emoji widget overlap the image picker.

The reason is that the "bottom" CSS value which is used to place the
widget when it's attached to a standard text field is not ignored by
brave when the parent DOM element has "display: flex".

Task-2291254

closes odoo/enterprise#12838

X-original-commit: 8afb734a6d225578677e2646639758779a34c04b
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>