[FIX] mrp_workorder: Add context info to start button
When return from the tablet view, we want to our previous view. If we
come from the tree view of the WO, we back to the tree view. If we come
from the workcenter kanban, we back to kanban view.
We add "from_production_order" to the context of the start button to
trace this info.

Task 2278147

X-original-commit: 2c06fbf9834911c00f3994c6d2428ec1e17d70a4