[FIX] social_{facebook, twitter}: fix bug after archiving the accounts
Bug
===
Archive a Facebook or a Twitter account, then open a social post in
the feed view. An error will be raised.

An error is also raised when we try to access to the social post Kanban
view, and if a social account has been archived (error when computing
`live_posts_by_media`).

Solution
========
Instead of calling `search_read`, we call `read` on the record itself,
so even if the record is archived, we can read its fields.

Task-2291254

X-original-commit 909d1e3a21dfda1e1cd540debc05380c030b3199

X-original-commit: 8116307d891baeec3ea3e5470f86c39e8e611d24