[FIX] l10n_be_hr_contract_salary: fix commission
Issue:
Have the commission advantages does not impact the wage
in the salary configurator. yearly_commission_cost is zero

This is due to a wrong configuration yearly_commission should not
be use as field for the advantage since it's a computed field and
the base of the computation is commission_on_target which is zero
since it's not display in the salary configurator

Solution:
Use commission_on_target in the salary configurator instead of
yearly_commission. Rename the method that display the dynamic
description

closes odoo/enterprise#11890

X-original-commit: a4e6e553a54907ce0267a39dc65e35b5bc3ea1b9
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>