[FIX] documents_spreasheet: Add columns according to subgroups
Insert a pivot in a spreadsheet where there are many values
in a subgroup such that the pivot takes more than the default 26 columns
in spreadsheet.
=> traceback

When inserting the pivot, computing the size of the sheet does not
correctly takes into account subgroups.

Before this commit, the size was computed (differently) at two places:

1) in the  controller before inserting (the faulty code here).
2) in the plugin itself when rebuilding a pivot.

This commit removes the duplicated computation from the controller and
let the plugin handle it in  all cases.

closes odoo/enterprise#11939

X-original-commit: 7f455e11e050edf4046f817b3923ba527f90b6b2
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Signed-off-by: lul-odoo <LucasLefevre@users.noreply.github.com>
Co-authored-by: fleodoo <fle@odoo.com>