[FIX] *: prevent crash on recent versions of libsass
Recent versions of libsass now crash on improper CSS that would
have silently failed in the past.

eg: :last-child() :first-child()
=> should not be followed by parenthesis

They throw an unclear error message like this one:

```
Error: Custom property values may not be empty.
        on line 4092:23 of /stdin
>>                     &:last-child() {
```

This commit fixes our CSS to prevent the error message from popping and
breaking the SCSS compilation.

Tested with the following versions of python-libsass (installed with
pip): 0.12, 0.17.0, 0.18.0, 0.19.0, 0.20.0, this fix is meant to be
retrocompatible.

Side note:
It also appears that Ubuntu packages do not ship the exact same version
of the underlying "libsass", it begins to fail with the ones available in
Ubuntu 20.04

X-original-commit: f1b4550161551ae3b5ccc2488d434f6b86532e96

----------------------------------------

This commit is also a side port of https://github.com/odoo/odoo/pull/54589

[FIX] web: tests: fix typos in :{last,first}-child

Follow-up of odoo/enterprise#11881, in order to eliminate all similar
typos.

The CSS pseudo-class selectors are spelled :first-child and
:last-child, and never take any argument, as opposed to
:nth-child(<nth>) for example.

Ref: https://developer.mozilla.org/docs/Web/CSS/Pseudo-classes

jQuery doesn't care and matches with or without the (), but CSS engines
don't, and now libsass SCSS compilation crashes due to the empty
argument list (cfr. opw-2299465)

Better avoid confusion and fix the typo everywhere.

closes odoo/enterprise#11953

X-original-commit: f40771350dc33dc5d97c62b2484206831220a716
Signed-off-by: Richard Mathot (rim) <rim@openerp.com>
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>
Co-authored-by: Richard Mathot <rim@odoo.com>
Co-authored-by: Olivier Dony <odo@odoo.com>