[FIX] timesheet_grid: Fix the quantity label for unit_amount on tree
Currently, method _apply_timesheet_label will add label on field
unit_amount with widget timesheet_uom but the widget on this
field is changed on tree view from timesheet_grid as
timesheet_uom_timer so it will not work when the module is
installed.

So in this commit, We already passed view_type argument in method
_apply_timesheet_label so remove the override of _fields_view_get
and apply same label if view_type is grid from _apply_timesheet_label.

Also check the OR condition for the widget timesheet_uom_timer so label
should be applied to tree view too.

closes odoo/enterprise#13279

Taskid: 2287390
Related: https://github.com/odoo/odoo/pull/54613
Related: odoo/odoo#57847
X-original-commit: 674c697a3ef7d4b880b1b0ffa96433751f3496a6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>