[FIX] mrp_workorder: fix backorder button showing
It does not make much sense to show the backorder when only the quality
checks are missing as you cant use the button if they are not done and
then once you do the quality checks the button backorder dissapear and
the button mark as done shows. Instead directly show the mark as done
button.

task-2278147

closes odoo/enterprise#11906

X-original-commit: 7443a6dfb904ec89436671cde52fde0629f23107
Related: odoo/odoo#54607
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>