[IMP] stock_barcode: put done lines at bottom
Done lines are now put at bottom of a [batch] picking's list and not
automatically scrolled to. Done lines do not include lines that are not
part of the original picking (as they can never be "Done"). It is also
assumed that it is OK a completed line can be placed at the very end of
the list (i.e. not in alphabetical order) while the user is still in the
view after completing the line. Done lines will be alphabetical after
user leaves the screen and returns.

Done lines are defined as having demain quantity completed and if has
serial number (SN)/lot tracking active then SN/lot is also assigned.
Done line does not take into account if package is assigned or not (when
setting is active).

Also add a background color to indicate that line is done.

Part of review tweaks for overall barcode improvement task.

Task: 1884520
Related COM PR: odoo/odoo#55096
Closes: odoo/enterprise#12086