[IMP] *: translations improvements
Avoid concatenation of translated strings & raw values and/or other
translated values, and use the new _ API (by providing the formatting
values to ensure the fallback on the english terms in case of wrong
translation).

By including the expected values inside the string to translate, we
ensure that the formatted values are included at the right place in the
resulting translated text.

This also provides more information on the string to translate to
the translators, s.t. the chosen terms for the translations can be more
adapted to the translation context.

closes odoo/enterprise#11925

Related: odoo/odoo#54635
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>