[IMP] stock_barcode: add barcodes to "commands for inventory"
This commit makes it so the "commands for Inventory" hyperlink in the
main menu of the barcode app includes operation types and location
barcodes in addition to the fixed commands. This is done by creating
the operation types/locations reports manually (since they are dynamic
depending on current company/active records) which are then appended
to the fixed commands pdf.

Only operation types/locations of the currently selected company that
have barcodes and are active will be included in the pdf.

Supports a review comment of overall barcode improvement task

Task: 1884520
COM PR: odoo/odoo#55096
Closes: odoo/enterprise#12086