[IMP] documents_*: Improve the setting usability
Purpose of this task is to improve document settings usability.

Currently, in the general settings of documents, there are some
fields that are not aligned with other fields. Also,in Payroll Workspace
default value is not set and there is no space between two fields.

In this commit, we set the default value of the payroll workshop and
space between two fields and align the fields. Alos prevent the dialog
on readonly tags list view.

closes odoo/enterprise#11994

Taskid: 2295969
Closes: #11994
Related: odoo/odoo#55255
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>