[FIX] sign: Fix the traceback on sample template
Currently, When the sample data template is deleted and when user click
on "start with our sample template" button will produce the traceback.

So in this commit, Hide the helper button when sample data template
is deleted after refresh but after the delete the helper button
is not replaced immediately as get_empty_list_help is not called so
On immediate click if template is not found then raise the warning
and return to kanban view.

closes odoo/enterprise#12224

Taskid: 2300897
Closes: #12098
X-original-commit: d10aa82e59789b8451e859752111f4e0156b2ad0
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>