[FIX] social_facebook: fir errors when we can not fetch comment author
Bug
===
Sometimes, we can not fetch the author name/picture/id of a comment.
In that case, when we try to open the comment, an error is raised.

Reproduce
1. Use a Facebook application in dev mode
2. With an user (who is not in the testing users in the Facebook application setting)
   comment the post on Facebook
3. Try to display it in Odoo

Solution
========
Instead of raising an error, we display "Unknown" as name with a default image.
So, we still can read the comment, even if we didn't know who is the author.

Task-2291254

X-original-commit 9450a28784eb5c6cc82d299b33e6696cc00e8a60

X-original-commit: 7cf087558ecb576f1486174feddd3f35c68a1004