[IMP] approvals: Improve the empty list help, new widget on list view
Purpose of the task is to use new widgets for the list views, improve
the empty screen help and improve usability of the apps.

So in this commit,
 - Use the new widget for approval list view
 - Improve the empty list view help message
 - Display approval category image to approval request
 - currently approval user should only delete the approval
which are in state 'cancel' but we have to allow user that he
can delete his own approval if it's in draft(yet to submit) and
cancelled.

closes odoo/enterprise#12167

Taskid: 2300320
Closes: #12167
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>