[IMP] stock_barcode: emphasize next expected scan action
This commit makes it so the next expected scan action has an associated
icon put in green as a visual cue to users. For now this means that for
all scan messages except for "scan_lot" the top line which is not "done"
(this excludes any lines not part of the original picking since they
can never be done) will have their fa-tag tagged with a
'o_next_expected' class. When "scan_lot" is the displayed message, then
the corresponding last scanned product that is tracked will have its
fa-barcode icon tagged. No icons are put in green when reservations are
completed in the current page.

This commit also rewords "Scan the source location" to "Scan the source
location, or scan a product" since it is not required to scan a source
location and user can start scanning products right away.

Supports a review comment of overall barcode improvement task

Task: 1884520
Related COM PR: odoo/odoo#55096
Closes: odoo/enterprise#12086