[FIX] account_reports, point_of_sale: make the tax report consider returns from POS as refunds when they don't generate an invoice
Before that, the tax report did not apply a -1 multiplier on the total amount for those move lines (even though the refund tags were properly assigned). We fix that by introducing a hook function into account_reports, and overriding it in a new pos_account_reports module, so that those lines are properply detected as coming from a credit note.

opw 2238358

closes odoo/enterprise#12320

X-original-commit: 1433f70b99a3653718117ce54f475012cbe37143
Related: odoo/odoo#55664
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>