[FIX] sale_subscription: perf issue on subscription portal
Problem:
the controller rely on record rules to select the payment.token
to be displayed on the portal's subscription page renewal.
This works fine with portal user, but internal user that
wants to preview, will face client side performance issue
as they can see all the token of the database

Solution:
Don't rely on record.rule in the controller. Use the domain
from the portal user rule in the search.

To make the search of token working for partners with more than
2 levels of hierachy, use child_of operator

closes odoo/enterprise#12579

X-original-commit: 4ce142c747ed6de43c3216eb038b13389adaaee8
Related: odoo/odoo#56326
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>