[IMP] documents: Imporve the emptyscreens and widget list
Purpose of this task is to update the help for emptyscreens
and list view widget for Documents App.

In this commit,
we show the badge in state filed of the document share tree view
and apply color accordingly als add the empty help content and
avatar_widget in documents.

closes odoo/enterprise#12117

Taskid: 2304292
Closes: #12117
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>