[FIX] l10n_mx_edi: Don't try to add addenda if it's empty
There are cases in which it is not necessary to add the addenda node so
the node should be empty.

eg: Fresko (out_refund invoices)

To support this case, an `if` statement is added in the addenda
template, which allows us to only render the addenda if the invoice
is "out_invoice"

Currently this causes an error, since this case is not supported by the
module.

To avoid this error, this change is added, where it is ensured to obtain
some value in the addenda node before trying to add it to the CFDI.

closes odoo/enterprise#12630

X-original-commit: 80e8c718f069cb72f0bb5ce53b8bb9a80e24807e
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>