[IMP] account_reports: add generic accounting report builder
Basically all you need to know is in the new class's docstring.

This class helps the creating of reports based on account.move.line by
creating a models.Model based on a VIEW and doing successive read_group
on the VIEW.
In most cases, this means we don't have to do custom optimisation an
custom layouts for the report. It also shrinks the size of each
implementation down by keeping only one SQL query and declaration of the
layout.