[FIX] hr_contract_report: fix 'Departure Employee' on employee report
Have an Employee with a contract end date day less than the start date
day
ex (start on 18/02/20, end on 08/04/20)
Reach the report Employees>Reporting>Contracts
The "Departure Employee" column should show a "1" for each employee that
have an active contract anymore.

This occur because we check the termination against a sequence with 'month'
stepping that will halt too soon when such condition is met.

Selectively adding an extra month fix the issue

opw-2272291

closes odoo/enterprise#12313

X-original-commit: dcef3e20f11ea8b7e775acb4a90690d802e468fe
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>