[FIX] Sign: avoid multiple control panel instances
Before this commit we were not waiting for `updateControlPanel()`.
Because of this the control panel was added twice after checking its
presence in owl_compatibility.js.

It seems that "isMounted" isn't correctly updated because we didn't
wait the promise.

On mobile some listeners (binded on scrolling event) are added twice by
mistake and these are not properly unmonted.

Errors then appear when events are not unbind correctly because some
DOM elements no longer exist after removing the control panel.

Steps to reproduce:

1. Go to 'Sign' and open a template
2. Go to 'Sales'
3. Scroll on the quotation view
=> Traceback from the control panel

closes odoo/enterprise#12239

X-original-commit: 748264c7064f96ecab1d4ce815bc0071a859ae09
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>