[FIX] web_enterprise: toggle navbar when editing a record
The problem arises when you edit any record and then click on the
home menu switcher. It displays the discard message but the navbar
is directly hidden and we already see the background image of the
home menu.

The issue occurred because the state of the webclient (whether the
home menu is displayed or not) was toggled directly, without
waiting for 'clear_uncommitted_changes' (i.e. without waiting for
the user to agree to discard changes).

Task : 2312437

closes odoo/enterprise#13107

X-original-commit: e667d3d40d0135f2e462274cb31a39a6cbe4debb
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>