[FIX] web: Correctly transmit fields to the control panel
Before this commit, the "fields" object given to the control panel was
altered by the view (e.g. dashboard populates it with its aggregates).
As a result the control panel and its subcomponents had an altered
object with more or less fields than expected.

Now, the initial "fields" object is given to avoid such problems.

Task 2312074

closes odoo/enterprise#12310

X-original-commit: b54f567a44784c0052faa7bddb0c3996aa2ad3c1
Related: odoo/odoo#55640
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>