[FIX] documents(_account): Fix the document Onboarding tour
Currenttly, the documents onborading tour is broken because of
data-id which is removed during searchpanel refactoring.

So in this commit, Fix the onboarding tour and use eq for
element selection instead of data-id and data-value-id as
both are removed from searchpanel and there is other specific
selector for particular element.

closes odoo/enterprise#12318

Taskid: 2313883
Closes: #12318
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>