[FIX] industry_fsm_sale: Traceback when editing product quantity
Currently, in the field services module when we edit the quantity
of a product by clicking on it and set blank then it produces a traceback.
Step to reproduce:
 -Create an FSM task
 -Products
 -edit the quantity of a product by clicking on it

So in this commit, if there is quantity in the product then it
it should check the condition. if quantity is blank then make
it 0.

closes odoo/enterprise#12789

Taskid: 2315103
Closes: #12330
X-original-commit: 1daa82bda82b4811334e1381657be53c09c50518
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>